Clippy, remove cache mod
